requirements.txt
PyQt5>=5.15.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.4.0
numpy>=1.21.0
reportlab>=3.6.0
openpyxl>=3.0.0
python-dateutil>=2.8.0
bcrypt>=4.1.0
